This script allows the bot to shoot projectiles of bullets out of it
Player Prefab edit: Added a firepoint as a child of the bot to indicate where the bullets will be coming from.
ProjectileShootingScript added: Will take the bullet as a game object, and will instantiate the bullet and shoot towards the direction of the mouse.

I have not added the damage of every shot and how it is going to affect enemies. I still need to read enemy script and understand so I can implement properly.